Loader: Remove .initMaterials() and .createMaterial().
